api/requirements.txt
annotated-types==0.7.0
anyio==4.9.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
distro==1.9.0
dnspython==2.7.0
email_validator==2.2.0
fastapi==0.115.12
fastapi-cli==0.0.7
greenlet==3.2.1
groq==0.23.0
h11==0.14.0
httpcore==1.0.8
httptools==0.6.4
httpx==0.28.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.3.24
langchain-core==0.3.55
langchain-groq==0.3.2
langchain-text-splitters==0.3.8
langsmith==0.3.33
lz==0.11.1
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
memoir==0.0.3
motor==3.7.0
mypy==0.991
mypy_extensions==1.1.0
numpy==2.2.5
orjson==3.10.16
packaging==24.2
pandas==2.2.3
paradigm==0.6.2
pydantic==2.11.3
pydantic-extra-types==2.10.3
pydantic-settings==2.9.1
pydantic_core==2.33.1
Pygments==2.19.1
pymongo==4.12.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-multipart==0.0.20
pytz==2025.2
PyYAML==6.0.2
redis==5.2.1
reprit==0.9.0
requests==2.32.3
requests-toolbelt==1.0.0
rich==14.0.0
rich-toolkit==0.14.1
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
spotipy==2.25.1
SQLAlchemy==2.0.40
starlette==0.46.2
tenacity==9.1.2
typer==0.15.2
typing-inspection==0.4.0
typing_extensions==4.13.2
tzdata==2025.2
ujson==5.10.0
urllib3==2.4.0
uvicorn==0.34.2
uvloop==0.21.0
watchfiles==1.0.5
websockets==15.0.1
zstandard==0.23.0